fix: resolve remaining 4 lint violations and fix test assertion case mismatch
- QF1008: remove embedded field "ServiceParam" from selector in write_buffer.go
- QF1008: remove embedded field "SearchRequest" from selector in task_search.go
- QF1003: convert if/else chain to switch statement in init_core.go
- unconvert: remove unnecessary int64 conversion in segment_manager_test.go
- Fix util_test.go assertion to match lowercase "unknown token_level" (ST1005)

+++ internal/datacoord/segment_manager_test.go
@@ -1126,7 +1126,7 @@ func TestAllocNewGrowingSegment_ManifestPath(t *testing.T) {
 		basePath, ver, unmarshalErr := packed.UnmarshalManifestPath(segment.ManifestPath)
 		assert.NoError(t, unmarshalErr)
 		assert.NotEmpty(t, basePath)
-		assert.Equal(t, int64(packed.ManifestEarliest), ver)
+		assert.Equal(t, packed.ManifestEarliest, ver)
 	})
 
 	t.Run("StorageV2 segment has no manifest path", func(t *testing.T) {
